Change WritingSystemRepository interaction with the Id field
* Stop treating the Canonicalized LanguageTag as a stable identifier
* Allow for multiple writing systems in the repository to have the same
  Canonicalized LanguageTag data
* Update the tag cleaner to handle some of the clean-up that
  Canonicalize did.
* Don't rename .ldml files if a writing system id has changed
  but is still equivalent
* Don't canonicalize language tags on WritingSystemDefinition creation

+++ SIL.WritingSystems/LdmlInFolderWritingSystemRepository.cs
@@ -145,9 +145,9 @@ namespace SIL.WritingSystems
 		///<summary>
 		/// Returns the full path to the underlying store for this writing system.
 		///</summary>
-		public string GetFilePathFromLanguageTag(string langTag)
+		public string GetFilePathFromLanguageTag(string wsId)
 		{
-			return Path.Combine(PathToWritingSystems, GetFileNameFromLanguageTag(langTag));
+			return Path.Combine(PathToWritingSystems, GetFileNameFromLanguageTag(wsId));
 		}
 
 		/// <summary>
@@ -280,7 +280,7 @@ namespace SIL.WritingSystems
 		{
 			Set(ws);
 
-			string writingSystemFilePath = GetFilePathFromLanguageTag(ws.LanguageTag);
+			string writingSystemFilePath = GetFilePathFromLanguageTag(ws.Id);
 			if (!File.Exists(writingSystemFilePath) && !string.IsNullOrEmpty(ws.Template) && File.Exists(ws.Template))
 			{
 				// this is a new writing system that was generated from a template, so copy the template over before saving
@@ -466,7 +466,7 @@ namespace SIL.WritingSystems
 			//helps us avoid having to deal with situations where a writing system id is changed to be
 			//identical with the old id of another writing sytsem. This could otherwise lead to dataloss.
 			//The inconsistency is resolved on Save()
-			if (oldStoreId != ws.Id && File.Exists(GetFilePathFromLanguageTag(oldStoreId)))
+			if (oldStoreId != ws.Id && File.Exists(GetFilePathFromLanguageTag(oldStoreId)) && !IetfLanguageTag.AreTagsEquivalent(oldStoreId, ws.Id))
 				File.Move(GetFilePathFromLanguageTag(oldStoreId), GetFilePathFromLanguageTag(ws.Id));
 		}
